+++ src/Changelog.ts
@@ -13,9 +13,15 @@ export default class Changelog {
   tagNameBuilder?: (release: Release) => string;
   /** @deprecated: Use tagLinkBuilder() instead */
   compareLinkBuilder?: (previous: Release, release: Release) => string;
-  tagLinkBuilder?: (url: string, tag: string, previous?: string, head?: string) => string;
+  tagLinkBuilder?: (
+    url: string,
+    tag: string,
+    previous?: string,
+    head?: string,
+  ) => string;
   format: "compact" | "markdownlint" = "compact";
   bulletStyle: "-" | "*" | "+" = "-";
+  autoSortReleases = true;
 
   constructor(title: string, description = "") {
     this.title = title;
@@ -24,7 +30,9 @@ export default class Changelog {
 
   addRelease(release: Release) {
     this.releases.push(release);
-    this.sortReleases();
+    if (this.autoSortReleases) {
+      this.sortReleases();
+    }
     release.changelog = this;
 
     return this;
@@ -54,11 +62,21 @@ export default class Changelog {
       const url = this.url!;
 
       if (!previous) {
-        return this.tagLinkBuilder(this.url!, this.tagName(release), undefined, this.head);
+        return this.tagLinkBuilder(
+          this.url!,
+          this.tagName(release),
+          undefined,
+          this.head,
+        );
       }
 
       if (!release.date || !release.version) {
-        return this.tagLinkBuilder(url, this.head, this.tagName(previous), this.head);
+        return this.tagLinkBuilder(
+          url,
+          this.head,
+          this.tagName(previous),
+          this.head,
+        );
       }
 
       return this.tagLinkBuilder(